Corrected bug in write_to_file function: now we save to the specified path instead of the project's folder.
